Spark is used as an ETL tool...
Created the sample XML data with the first 100,000 lines from
StackOverflow data dump file Users.xml.

StackOverflow Users.xml has 1.8 GB and 5987287 lines as of September
2016.

Loaded the XML data into a DataFrame and queried it with Spark SQL.
Later the query result is saved to a CSV file to be processed by the
spark-frontend [ HTML + JavaScript ] app.

The plan is to plot some nice charts with the help of D3.js