Added code for spark-frontend using PlayFramework for Scala https://www.playframework.com/
All the CSV files generated by spark-backend are available in
spark-frontend. A listing is presented to the user where they can click
the link and download the CSV to further process\analyze\massage the
data in spreadsheet software.
 
Changed the location of the generated JAR file for spark-backend so that
it's possible to share the input folder when submitting the jar with
spark-submit.